backend/requirements.txt
--index-url https://pypi.org/simple
--extra-index-url https://download.pytorch.org/whl/cpu
torch>=2.1.0
torchvision>=0.16.0
fastapi
uvicorn
python-multipart
transformers
pillow
requests
numpy